Move simulation-related things to sim.cpp
